add progress callback to encryptProgressUploadSearchableBlob
this adds a callback which will be called with the a number between 0
and 1 representing the percentage of the file that has been encrypted
and uploaded to minio.

The number will represent the percentage complete in relation the
previously uploaded fragment, not the current fragment. And therefore
will always be <1 in the inner loop of encryptProgressBlob. The callback
is called manually with a value of 1 after the _part and _meta files
have been uploaded, and the uploadData function has been called.